backend/requirements.txt
# FastAPI and Server Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.0
email-validator==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# ML/AI Dependencies
torch==2.2.0
transformers==4.37.0
spacy==3.7.2
nltk==3.8.1
stable-baselines3==2.2.1
openai==1.12.0

# Data Processing
pandas==2.2.0
numpy==1.26.0
python-docx==1.1.0
PyPDF2==3.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

# Development Tools
python-dotenv==1.0.0
black==24.1.1
flake8==7.0.0
isort==5.13.2 